Split C comms template into files, add Makefile
