+++ Makefile
@@ -1,0 +1,32 @@
+# Compiler
+CC = gcc
+
+# Compiler flags
+CFLAGS = -Wall -Wextra -O2
+
+# Target executable
+TARGET = idle_communication_template
+
+# Source files
+SRCS = idle_communication_template.c functions.c
+
+# Object files
+OBJS = $(SRCS:.c=.o)
+
+# Default target
+all: $(TARGET)
+
+# Build the target executable
+$(TARGET): $(OBJS)
+	$(CC) $(CFLAGS) -o $@ $(OBJS)
+
+# Compile source files to object files
+%.o: %.c
+	$(CC) $(CFLAGS) -c $< -o $@
+
+# Clean up
+clean:
+	rm -f $(TARGET) $(OBJS)
+
+# Phony targets
+.PHONY: all clean
